+++ package/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1
+
+* An `expand` property has been added to `EasyLayoutRow` to control how items are filled in a row. 
+* **Breaking change:** `EasyLayoutFlexible.flex` is now `int` instead of `double`.
+
 ## 1.0.0+1
 
 * EasyLayout now inherits spacing from parent EasyLayout